LLM Gotcha Riddles (#85)
**Describe the contribution to the taxonomy**

This attemps to give the LLM an advantage when asked some "gotcha"
riddle type questions that often confuse LLMs. I saw both of these in
the video "A Hackers' Guide to Language Models" by Jeremy Howard. I plan
on adding additional riddles as I find them but this gives the general
idea of the answer that needs to be given for a given riddle.

**Input given at the prompt (Riddle 1)**
Solve this riddle: I'm in my house. On top of my chair in the living
room is a coffee cup. Inside the coffee cup is a thimble. Inside the
thimble is a single diamond. I move the chair to my bedroom. Then I put
the coffee cup on the bed. Then I turn the cup upside down. Then I
return it to right-side-up, and place the coffee cup on the kitchen
counter. Where is my diamond?

**Response that was received (Riddle 1)**
The diamond is still inside the thimble. The riddle mentions that you
move the chair to your bedroom, put the coffee cup on the bed, turn it
upside down, and then place the coffee cup on the kitchen counter.
However, the actions you perform on the coffee cup do not affect the
position of the diamond, which remains inside the thimble throughout the
process.

**Response that is now received instead (Riddle 1)**
This riddle presents a series of actions that involve moving a coffee
cup with a thimble containing a diamond from the living room to the
kitchen counter. To determine the final location of the diamond, we need
to analyze the sequence of actions step-by-step:

1. The diamond is initially placed inside the thimble, which is inside
the coffee cup, on the chair in the living room.
2. The chair is moved to the bedroom, but the coffee cup remains on the
bed.
3. The coffee cup is turned upside down, causing the thimble (and the
diamond) to fall out.
4. The coffee cup is returned to its right-side-up position on the
kitchen counter.

Since the thimble fell out when the coffee cup was turned upside down,
the diamond is no longer inside the thimble or the coffee cup.
Therefore, the diamond is now on the bed, having fallen out from the
thimble when the coffee cup was turned upside down.

**Input given at the prompt (Riddle 2)**
Sally (a girl) has 3 brothers. Each brother has 2 sisters. How many
sisters does Sally have?

**Response that was received (Riddle 2)**
Sally is one of the brothers' sisters, so she counts as one sister. Each
of her two other brothers share the same two sisters, so the other two
sisters are counted twice. Therefore, Sally has two sisters in total.

**Response that is now received instead (Riddle 2)**
This is actually a trick question! Let's break down the given
information:

1. Sally is one of the girls.
2. There are three boys, including Sally's brothers.

Now, let's analyze the statement "Each brother has 2 sisters." This
means that each of the three brothers has two sisters, but it does not
necessarily mean that all of the sisters are different individuals. In
fact, it is possible that one sister is shared among the brothers, which
is Sally herself.

So, if we consider the shared sister, the answer to the question would
be:

Sally has 1 sister (her younger sister, for example). The other two
siblings mentioned in the question are her three brothers.

It's essential to be aware of the pitfalls in such questions, as they
might not always have straightforward answers.

**Please, confirm that you have tested your contribution with `lab
generate` and ensured that it does not produce any warnings or errors:**

Yes
```
lab generate
INFO 2024-03-05 17:14:44,362 lab.py:275 Generating model 'ggml-merlinite-7b-0302-Q4_K_M' using 10 cpus, taxonomy: 'taxonomy' and seed 'seed_tasks.json'
  0%|                                                                                                                                                                                       | 0/100 [00:00<?, ?it/s]Cannot find prompt.txt. Using default prompt.
100%|█████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 100/100 [25:14<00:00, 21.84s/it]INFO 2024-03-05 17:39:59,385 generate_data.py:444 Generation took 1515.02s
100%|█████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 100/100 [25:14<00:00, 15.15s/it]
```

Signed-off-by: Dave Wilde <dwilde@redhat.com>

Signed-off-by: Dave Wilde <dwilde@redhat.com>