Add skill to count punctuation marks in text (#89)
**Describe the contribution to the taxonomy**

Add a skill to count the total number of punctuation marks present in a
given piece of text.


**Input given at the prompt**

Count the total number of punctuation marks in the following text:
Hello, world! How are you?


**Response that was received**

To count the total number of punctuation marks in the given text
\"Hello, world! How are you?\", we first need to identify what
punctuation marks are present. The punctuation marks in this text are:
comma (,), exclamation mark (!), question mark (?), and space ( ). Since
there is one exclamation mark and one question mark, there are 2
punctuation marks in the given text.


**Response that is now received instead**

The text has 3 punctuation marks in total


**Please, confirm that you have tested your contribution with `lab
generate` and ensured that it does not produce any warnings or errors:**

Yes

Signed-off-by: Mohit Gaur <56885276+Mohit-Gaur@users.noreply.github.com>