Add technical_skills and time calucations (#101)
Add technical_skills and seed with some time calucations

**Describe the contribution to the taxonomy**

Adding ability to calculate hours between times of the day


**Input given at the prompt**
>>> How many hours between 10pm and 11pm the following day

**Response that was received**
To find out how many hours there are between 10 pm and 11 pm the next
day, we first need to understand that there are 24 hours in a day. 10 pm
is considered as part of the previous day, so let's break it down:

 - From 10 pm to midnight, there is 1 hour.
- From midnight to 1 am, there is another hour (even though it's
technically still the same day).
 - From 1 am to 2 am, there is another hour.
 - From 2 am to 3 am, there is another hour.
 - From 3 am to 4 am, there is another hour.
 - From 4 am to 5 am, there is another hour.
 - From 5 am to 6 am, there is another hour.
 - From 6 am to 7 am, there is another hour.
 - From 7 am to 8 am, there is another hour.
 - From 8 am to 9 am, there is another hour.
 - From 9 am to 10 am, there is another hour.

 Adding these up, there are 13 hours from 10 pm to 11 pm the next day.

**Input given at the prompt**
>>> hour many hours from 10PM to 2AM

**Response that was received**
To find out how many hours there are between 10 pm and 2 am, we first
need to understand that there are 24 hours in a day. 10 pm is considered
as part of the previous day, so let's break it down:

 - From 10 pm to 11 pm, there is 1 hour.
- From 11 pm to midnight, there is another hour (even though it's
technically still the same day).
- From midnight to 1 am, there is another hour (since we're moving from
one day to the next).

 Adding these up, there are 3 hours from 10 pm to 2 am.


**Response that is now received instead**
Not retrained

**Please, confirm that you have tested your contribution with `lab
generate` and ensured that it does not produce any warnings or errors:**

Yes

Signed-off-by: Derek Higgins <derekh@redhat.com>