Create a Linux development skill to determine userspace memory page size (#149)
* Extract Linux userspace memory page size from a given context.

Explains the skill of extracting a userspace memory page size from a
given context.

The existing answers from the model are somewhat poor for the dynamic
loader auxiliary vector diagnostic information. The best answer is had
when extracting from getconf output, but the skill provides 3 ways to
perform the task (proc, cli, ld.so) to try to link the other two better
answers.

**Describe the contribution to the taxonomy**

- Adds a Linux development skill to determine the userspace memory page
size.

The userspace memory page size is important to developers for optimizing
placement of data in userspace applications.

**Input given at the prompt**

<!-- What you entered, replace "..." -->

```
>>> Extract the userspace memory page size given the context: ...
```
With the context for extraction provided.

**Response that was received**

<!-- What you received in response to your input, replace "..." -->

```
╭──────────────────────────────────────────── ggml-merlinite-7b-Q4_K_M ────────────────────────────────────────────╮
│ The Linux kernel page size can be determined from this text using the `/usr/bin/ld.so --list-diagnostics`        │
│ command followed by `grep auxv`. The command displays various auxiliary vector (auxv) entries, one of which is   │
│ for the page size. The `a_type` field of 0x10 corresponds to the page size, and its `a_val` field is 0xbfebfbff, │
│ which is 4096 in decimal, representing 4KiB page size.                                                           │
╰───────────────────────────────────────────────────────────────────────────────────────── elapsed 27.806 seconds ─╯
>
```
The response is wrong in both the type and value, but the answer is
correct and we should improve that.

**Response that is now received instead**

Did not run chat with a new model on x86_64 Linux laptop. Only ran
``yamllint`` to verify and ``lab generate`` to ensure generation works.

The answer given by the model is poor in this case and I expect my skill
to determine page size will improve the skill.

The answer is better for ``getconf``, but I want to improve the
diagnostic extraction so I use ``getconf`` in my Q&A help link together
the same concept.

```
╭──────────────────────────────────────────── ggml-merlinite-7b-Q4_K_M ────────────────────────────────────────────╮
│ The `getconf` command is used to retrieve various system configuration details, including the userspace memory   │
│ page size. By executing `getconf PAGE_SIZE`, we can determine the current page size in userspace. In this case,  │
│ the output is 4096, which means the userspace memory page size is 4KiB.                                          │
╰───────────────────────────────────────────────────────────────────────────────────────── elapsed 17.471 seconds ─╯
```
Likewise I include the `procfs` example to similarly link the concept:
```
╭──────────────────────────────────────────── ggml-merlinite-7b-Q4_K_M ────────────────────────────────────────────╮
│ The `/proc/self/smaps` file contains memory mapping information about the current process. By grepping for       │
│ `KernelPageSize` in this file, we can extract the userspace memory page size. The `head -1` command is used to   │
│ retrieve only the first line of output. In this case, the output indicates that the userspace memory page size   │
│ is 4 kB.                                                                                                         │
╰───────────────────────────────────────────────────────────────────────────────────────── elapsed 19.936 seconds ─╯
```
The answer here could be improved to say ``4KiB`` since the kernel means
that but doesn't print it.


**Contribution checklist**

- [x] tested contribution with `lab generate`
- [x] `lab generate` does not produce any warnings or errors
- [x] all [commits are signed
off](https://github.com/instruct-lab/taxonomy/blob/main/CONTRIBUTING.md#legal)
(DCO)
- [x] the `qna.yaml` file was [linted](https://yamllint.com)

Signed-off-by: Carlos O'Donell <carlos@redhat.com>