Improve word frequency count skill (#244)
Improve word frequency count skill by adding an additional example of
such a task.

If your PR is related to a contribution to the taxonomy, please, fill
out the following questionnaire. If not, replace this whole text and the
following questionnaire with whatever information is applicable to your
PR.


**Describe the contribution to the taxonomy**

<!-- A concise description of what the contribution brings, replace
"..." in the bullet list -->

- Improve on counting frequency of words in a sentence


**Input given at the prompt**

<!-- What you entered, replace "..." -->

```
Count the frequency of words in the following text and provide the result in the format [(word, frequency), (word,frequency), ...]: "A word is considered to be high frequency if
the word is commonly used in daily speech, such as the word "the". A word is considered to be low frequency if the word is not commonly used, such as the word "strait". Some language
s such as Chinese have multiple levels of daily speech that impact frequency of words."
```


**Response that was received**

<!-- What you received in response to your input, replace "..." -->

```
╭────────────────────────────────────────────────────────────────────────── ggml-merlinite-7b-0302-Q4_K_M ───────────────────────────────────────────────────────────────────────────╮
│ [('the', 1), ('word', 1), ('is', 2), ('commonly', 1), ('used', 1), ('daily', 1), ('speech', 1), ('such', 1), ('as', 1), ('the', 1), ('word', 1), ('strait', 1), ('languages', 1),  │
│ ('such', 1), ('as', 1), ('Chinese', 1), ('have', 1), ('multiple', 1), ('levels', 1), ('of', 1), ('daily', 1), ('speech', 1), ('that', 1), ('impact', 1), ('frequency', 1), ('of',  │
│ 1), ('words', 1)]                                                                                                                                                                  │
╰─────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────── elapsed 45.463 seconds ─╯
```


**Response that is now received instead**

<!-- What you receive with your contribution, replace "..." -->

```
Not applicable as I wasn't yet able to re-train the model.
```

**Contribution checklist**

<!-- Insert an x between the empty brackets: [ ] >> [x] -->

- [x ] tested contribution with `lab generate`
- [x ] `lab generate` does not produce any warnings or errors
- [x ] all [commits are signed
off](https://github.com/instruct-lab/taxonomy/blob/main/CONTRIBUTING.md#legal)
(DCO)
- [x ] the `qna.yaml` file was [linted](https://yamllint.com)

Signed-off-by: Juri Lelli <juri.lelli@redhat.com>