Give me an example of an Irish proverb (#246)
If your PR is related to a contribution to the taxonomy, please, fill
out the following questionnaire. If not, replace this whole text and the
following questionnaire with whatever information is applicable to your
PR.

Describe the contribution to the taxonomy

adding Proverb folder
adding Irish proverb
Input given at the prompt

Give me an example of an Irish proverb about finishing a task

Response that was received
An Irish proverb that highlights the importance of finishing what you
start is "A good beginning makes a good ending." This proverb emphasizes
the significance of taking the first steps towards completing a task, as
it often paves the way for a successful conclusion. By investing time,
effort, and dedication in the initial stages, one can create a solid
foundation that will ultimately lead to the completion of the project or
goal.

Response that is now received instead
Ní neart go cur le chéile. Translation: "There is no strength without
unity." This proverb can also be interpreted as emphasizing the
importance of teamwork and collaboration in getting work done, as it
takes the combined effort and determination of a team to accomplish
great things.

Contribution checklist

[ x] tested contribution with lab generate
[ x] lab generate does not produce any warnings or errors
[x ] all [commits are signed
off](https://github.com/instruct-lab/taxonomy/blob/main/CONTRIBUTING.md#legal)
(DCO)
[x ] the qna.yaml file was [linted](https://yamllint.com/)

---------

Signed-off-by: Judy Kelly <jkelly@redhat.com>
Co-authored-by: Christian Kadner <ckadner@us.ibm.com>